Module aws_sdk_greengrass::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AssociateRoleToGroup

Fluent builder constructing a request to AssociateRoleToGroup.

AssociateServiceRoleToAccount

Fluent builder constructing a request to AssociateServiceRoleToAccount.

CreateConnectorDefinition

Fluent builder constructing a request to CreateConnectorDefinition.

CreateConnectorDefinitionVersion

Fluent builder constructing a request to CreateConnectorDefinitionVersion.

CreateCoreDefinition

Fluent builder constructing a request to CreateCoreDefinition.

CreateCoreDefinitionVersion

Fluent builder constructing a request to CreateCoreDefinitionVersion.

CreateDeployment

Fluent builder constructing a request to CreateDeployment.

CreateDeviceDefinition

Fluent builder constructing a request to CreateDeviceDefinition.

CreateDeviceDefinitionVersion

Fluent builder constructing a request to CreateDeviceDefinitionVersion.

CreateFunctionDefinition

Fluent builder constructing a request to CreateFunctionDefinition.

CreateFunctionDefinitionVersion

Fluent builder constructing a request to CreateFunctionDefinitionVersion.

CreateGroup

Fluent builder constructing a request to CreateGroup.

CreateGroupCertificateAuthority

Fluent builder constructing a request to CreateGroupCertificateAuthority.

CreateGroupVersion

Fluent builder constructing a request to CreateGroupVersion.

CreateLoggerDefinition

Fluent builder constructing a request to CreateLoggerDefinition.

CreateLoggerDefinitionVersion

Fluent builder constructing a request to CreateLoggerDefinitionVersion.

CreateResourceDefinition

Fluent builder constructing a request to CreateResourceDefinition.

CreateResourceDefinitionVersion

Fluent builder constructing a request to CreateResourceDefinitionVersion.

CreateSoftwareUpdateJob

Fluent builder constructing a request to CreateSoftwareUpdateJob.

CreateSubscriptionDefinition

Fluent builder constructing a request to CreateSubscriptionDefinition.

CreateSubscriptionDefinitionVersion

Fluent builder constructing a request to CreateSubscriptionDefinitionVersion.

DeleteConnectorDefinition

Fluent builder constructing a request to DeleteConnectorDefinition.

DeleteCoreDefinition

Fluent builder constructing a request to DeleteCoreDefinition.

DeleteDeviceDefinition

Fluent builder constructing a request to DeleteDeviceDefinition.

DeleteFunctionDefinition

Fluent builder constructing a request to DeleteFunctionDefinition.

DeleteGroup

Fluent builder constructing a request to DeleteGroup.

DeleteLoggerDefinition

Fluent builder constructing a request to DeleteLoggerDefinition.

DeleteResourceDefinition

Fluent builder constructing a request to DeleteResourceDefinition.

DeleteSubscriptionDefinition

Fluent builder constructing a request to DeleteSubscriptionDefinition.

DisassociateRoleFromGroup

Fluent builder constructing a request to DisassociateRoleFromGroup.

DisassociateServiceRoleFromAccount

Fluent builder constructing a request to DisassociateServiceRoleFromAccount.

GetAssociatedRole

Fluent builder constructing a request to GetAssociatedRole.

GetBulkDeploymentStatus

Fluent builder constructing a request to GetBulkDeploymentStatus.

GetConnectivityInfo

Fluent builder constructing a request to GetConnectivityInfo.

GetConnectorDefinition

Fluent builder constructing a request to GetConnectorDefinition.

GetConnectorDefinitionVersion

Fluent builder constructing a request to GetConnectorDefinitionVersion.

GetCoreDefinition

Fluent builder constructing a request to GetCoreDefinition.

GetCoreDefinitionVersion

Fluent builder constructing a request to GetCoreDefinitionVersion.

GetDeploymentStatus

Fluent builder constructing a request to GetDeploymentStatus.

GetDeviceDefinition

Fluent builder constructing a request to GetDeviceDefinition.

GetDeviceDefinitionVersion

Fluent builder constructing a request to GetDeviceDefinitionVersion.

GetFunctionDefinition

Fluent builder constructing a request to GetFunctionDefinition.

GetFunctionDefinitionVersion

Fluent builder constructing a request to GetFunctionDefinitionVersion.

GetGroup

Fluent builder constructing a request to GetGroup.

GetGroupCertificateAuthority

Fluent builder constructing a request to GetGroupCertificateAuthority.

GetGroupCertificateConfiguration

Fluent builder constructing a request to GetGroupCertificateConfiguration.

GetGroupVersion

Fluent builder constructing a request to GetGroupVersion.

GetLoggerDefinition

Fluent builder constructing a request to GetLoggerDefinition.

GetLoggerDefinitionVersion

Fluent builder constructing a request to GetLoggerDefinitionVersion.

GetResourceDefinition

Fluent builder constructing a request to GetResourceDefinition.

GetResourceDefinitionVersion

Fluent builder constructing a request to GetResourceDefinitionVersion.

GetServiceRoleForAccount

Fluent builder constructing a request to GetServiceRoleForAccount.

GetSubscriptionDefinition

Fluent builder constructing a request to GetSubscriptionDefinition.

GetSubscriptionDefinitionVersion

Fluent builder constructing a request to GetSubscriptionDefinitionVersion.

GetThingRuntimeConfiguration

Fluent builder constructing a request to GetThingRuntimeConfiguration.

ListBulkDeploymentDetailedReports

Fluent builder constructing a request to ListBulkDeploymentDetailedReports.

ListBulkDeployments

Fluent builder constructing a request to ListBulkDeployments.

ListConnectorDefinitionVersions

Fluent builder constructing a request to ListConnectorDefinitionVersions.

ListConnectorDefinitions

Fluent builder constructing a request to ListConnectorDefinitions.

ListCoreDefinitionVersions

Fluent builder constructing a request to ListCoreDefinitionVersions.

ListCoreDefinitions

Fluent builder constructing a request to ListCoreDefinitions.

ListDeployments

Fluent builder constructing a request to ListDeployments.

ListDeviceDefinitionVersions

Fluent builder constructing a request to ListDeviceDefinitionVersions.

ListDeviceDefinitions

Fluent builder constructing a request to ListDeviceDefinitions.

ListFunctionDefinitionVersions

Fluent builder constructing a request to ListFunctionDefinitionVersions.

ListFunctionDefinitions

Fluent builder constructing a request to ListFunctionDefinitions.

ListGroupCertificateAuthorities

Fluent builder constructing a request to ListGroupCertificateAuthorities.

ListGroupVersions

Fluent builder constructing a request to ListGroupVersions.

ListGroups

Fluent builder constructing a request to ListGroups.

ListLoggerDefinitionVersions

Fluent builder constructing a request to ListLoggerDefinitionVersions.

ListLoggerDefinitions

Fluent builder constructing a request to ListLoggerDefinitions.

ListResourceDefinitionVersions

Fluent builder constructing a request to ListResourceDefinitionVersions.

ListResourceDefinitions

Fluent builder constructing a request to ListResourceDefinitions.

ListSubscriptionDefinitionVersions

Fluent builder constructing a request to ListSubscriptionDefinitionVersions.

ListSubscriptionDefinitions

Fluent builder constructing a request to ListSubscriptionDefinitions.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

ResetDeployments

Fluent builder constructing a request to ResetDeployments.

StartBulkDeployment

Fluent builder constructing a request to StartBulkDeployment.

StopBulkDeployment

Fluent builder constructing a request to StopBulkDeployment.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateConnectivityInfo

Fluent builder constructing a request to UpdateConnectivityInfo.

UpdateConnectorDefinition

Fluent builder constructing a request to UpdateConnectorDefinition.

UpdateCoreDefinition

Fluent builder constructing a request to UpdateCoreDefinition.

UpdateDeviceDefinition

Fluent builder constructing a request to UpdateDeviceDefinition.

UpdateFunctionDefinition

Fluent builder constructing a request to UpdateFunctionDefinition.

UpdateGroup

Fluent builder constructing a request to UpdateGroup.

UpdateGroupCertificateConfiguration

Fluent builder constructing a request to UpdateGroupCertificateConfiguration.

UpdateLoggerDefinition

Fluent builder constructing a request to UpdateLoggerDefinition.

UpdateResourceDefinition

Fluent builder constructing a request to UpdateResourceDefinition.

UpdateSubscriptionDefinition

Fluent builder constructing a request to UpdateSubscriptionDefinition.

UpdateThingRuntimeConfiguration

Fluent builder constructing a request to UpdateThingRuntimeConfiguration.