Struct aws_sdk_greengrass::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for Greengrass.

This client allows ergonomic access to a Greengrass-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_role_to_group(&self) -> AssociateRoleToGroup<C, M, R>

pub fn associate_service_role_to_account(
&self
) -> AssociateServiceRoleToAccount<C, M, R>

pub fn create_connector_definition(&self) -> CreateConnectorDefinition<C, M, R>

pub fn create_connector_definition_version(
&self
) -> CreateConnectorDefinitionVersion<C, M, R>

pub fn create_core_definition(&self) -> CreateCoreDefinition<C, M, R>

pub fn create_core_definition_version(
&self
) -> CreateCoreDefinitionVersion<C, M, R>

pub fn create_deployment(&self) -> CreateDeployment<C, M, R>

pub fn create_device_definition(&self) -> CreateDeviceDefinition<C, M, R>

pub fn create_device_definition_version(
&self
) -> CreateDeviceDefinitionVersion<C, M, R>

pub fn create_function_definition(&self) -> CreateFunctionDefinition<C, M, R>

pub fn create_function_definition_version(
&self
) -> CreateFunctionDefinitionVersion<C, M, R>

pub fn create_group(&self) -> CreateGroup<C, M, R>

pub fn create_group_certificate_authority(
&self
) -> CreateGroupCertificateAuthority<C, M, R>

pub fn create_group_version(&self) -> CreateGroupVersion<C, M, R>

pub fn create_logger_definition(&self) -> CreateLoggerDefinition<C, M, R>

pub fn create_logger_definition_version(
&self
) -> CreateLoggerDefinitionVersion<C, M, R>

pub fn create_resource_definition(&self) -> CreateResourceDefinition<C, M, R>

pub fn create_resource_definition_version(
&self
) -> CreateResourceDefinitionVersion<C, M, R>

pub fn create_software_update_job(&self) -> CreateSoftwareUpdateJob<C, M, R>

pub fn create_subscription_definition(
&self
) -> CreateSubscriptionDefinition<C, M, R>

pub fn create_subscription_definition_version(
&self
) -> CreateSubscriptionDefinitionVersion<C, M, R>

pub fn delete_connector_definition(&self) -> DeleteConnectorDefinition<C, M, R>

pub fn delete_core_definition(&self) -> DeleteCoreDefinition<C, M, R>

pub fn delete_device_definition(&self) -> DeleteDeviceDefinition<C, M, R>

pub fn delete_function_definition(&self) -> DeleteFunctionDefinition<C, M, R>

pub fn delete_group(&self) -> DeleteGroup<C, M, R>

pub fn delete_logger_definition(&self) -> DeleteLoggerDefinition<C, M, R>

pub fn delete_resource_definition(&self) -> DeleteResourceDefinition<C, M, R>

pub fn delete_subscription_definition(
&self
) -> DeleteSubscriptionDefinition<C, M, R>

pub fn disassociate_role_from_group(&self) -> DisassociateRoleFromGroup<C, M, R>

pub fn disassociate_service_role_from_account(
&self
) -> DisassociateServiceRoleFromAccount<C, M, R>

pub fn get_associated_role(&self) -> GetAssociatedRole<C, M, R>

pub fn get_bulk_deployment_status(&self) -> GetBulkDeploymentStatus<C, M, R>

pub fn get_connectivity_info(&self) -> GetConnectivityInfo<C, M, R>

pub fn get_connector_definition(&self) -> GetConnectorDefinition<C, M, R>

pub fn get_connector_definition_version(
&self
) -> GetConnectorDefinitionVersion<C, M, R>

pub fn get_core_definition(&self) -> GetCoreDefinition<C, M, R>

pub fn get_core_definition_version(&self) -> GetCoreDefinitionVersion<C, M, R>

pub fn get_deployment_status(&self) -> GetDeploymentStatus<C, M, R>

pub fn get_device_definition(&self) -> GetDeviceDefinition<C, M, R>

pub fn get_device_definition_version(
&self
) -> GetDeviceDefinitionVersion<C, M, R>

pub fn get_function_definition(&self) -> GetFunctionDefinition<C, M, R>

pub fn get_function_definition_version(
&self
) -> GetFunctionDefinitionVersion<C, M, R>

pub fn get_group(&self) -> GetGroup<C, M, R>

pub fn get_group_certificate_authority(
&self
) -> GetGroupCertificateAuthority<C, M, R>

pub fn get_group_certificate_configuration(
&self
) -> GetGroupCertificateConfiguration<C, M, R>

pub fn get_group_version(&self) -> GetGroupVersion<C, M, R>

pub fn get_logger_definition(&self) -> GetLoggerDefinition<C, M, R>

pub fn get_logger_definition_version(
&self
) -> GetLoggerDefinitionVersion<C, M, R>

pub fn get_resource_definition(&self) -> GetResourceDefinition<C, M, R>

pub fn get_resource_definition_version(
&self
) -> GetResourceDefinitionVersion<C, M, R>

pub fn get_service_role_for_account(&self) -> GetServiceRoleForAccount<C, M, R>

pub fn get_subscription_definition(&self) -> GetSubscriptionDefinition<C, M, R>

pub fn get_subscription_definition_version(
&self
) -> GetSubscriptionDefinitionVersion<C, M, R>

pub fn get_thing_runtime_configuration(
&self
) -> GetThingRuntimeConfiguration<C, M, R>

pub fn list_bulk_deployment_detailed_reports(
&self
) -> ListBulkDeploymentDetailedReports<C, M, R>

pub fn list_bulk_deployments(&self) -> ListBulkDeployments<C, M, R>

pub fn list_connector_definitions(&self) -> ListConnectorDefinitions<C, M, R>

pub fn list_connector_definition_versions(
&self
) -> ListConnectorDefinitionVersions<C, M, R>

pub fn list_core_definitions(&self) -> ListCoreDefinitions<C, M, R>

pub fn list_core_definition_versions(
&self
) -> ListCoreDefinitionVersions<C, M, R>

pub fn list_deployments(&self) -> ListDeployments<C, M, R>

pub fn list_device_definitions(&self) -> ListDeviceDefinitions<C, M, R>

pub fn list_device_definition_versions(
&self
) -> ListDeviceDefinitionVersions<C, M, R>

pub fn list_function_definitions(&self) -> ListFunctionDefinitions<C, M, R>

pub fn list_function_definition_versions(
&self
) -> ListFunctionDefinitionVersions<C, M, R>

pub fn list_group_certificate_authorities(
&self
) -> ListGroupCertificateAuthorities<C, M, R>

pub fn list_groups(&self) -> ListGroups<C, M, R>

pub fn list_group_versions(&self) -> ListGroupVersions<C, M, R>

pub fn list_logger_definitions(&self) -> ListLoggerDefinitions<C, M, R>

pub fn list_logger_definition_versions(
&self
) -> ListLoggerDefinitionVersions<C, M, R>

pub fn list_resource_definitions(&self) -> ListResourceDefinitions<C, M, R>

pub fn list_resource_definition_versions(
&self
) -> ListResourceDefinitionVersions<C, M, R>

pub fn list_subscription_definitions(
&self
) -> ListSubscriptionDefinitions<C, M, R>

pub fn list_subscription_definition_versions(
&self
) -> ListSubscriptionDefinitionVersions<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn reset_deployments(&self) -> ResetDeployments<C, M, R>

pub fn start_bulk_deployment(&self) -> StartBulkDeployment<C, M, R>

pub fn stop_bulk_deployment(&self) -> StopBulkDeployment<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_connectivity_info(&self) -> UpdateConnectivityInfo<C, M, R>

pub fn update_connector_definition(&self) -> UpdateConnectorDefinition<C, M, R>

pub fn update_core_definition(&self) -> UpdateCoreDefinition<C, M, R>

pub fn update_device_definition(&self) -> UpdateDeviceDefinition<C, M, R>

pub fn update_function_definition(&self) -> UpdateFunctionDefinition<C, M, R>

pub fn update_group(&self) -> UpdateGroup<C, M, R>

pub fn update_group_certificate_configuration(
&self
) -> UpdateGroupCertificateConfiguration<C, M, R>

pub fn update_logger_definition(&self) -> UpdateLoggerDefinition<C, M, R>

pub fn update_resource_definition(&self) -> UpdateResourceDefinition<C, M, R>

pub fn update_subscription_definition(
&self
) -> UpdateSubscriptionDefinition<C, M, R>

pub fn update_thing_runtime_configuration(
&self
) -> UpdateThingRuntimeConfiguration<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self