Module aws_sdk_greengrass::model

Modules

bulk_deployment

See BulkDeployment

bulk_deployment_metrics

See BulkDeploymentMetrics

bulk_deployment_result

See BulkDeploymentResult

connectivity_info

See ConnectivityInfo

connector

See Connector

connector_definition_version

See ConnectorDefinitionVersion

core

See Core

core_definition_version

See CoreDefinitionVersion

definition_information

See DefinitionInformation

deployment

See Deployment

device

See Device

device_definition_version

See DeviceDefinitionVersion

error_detail

See ErrorDetail

function

See Function

function_configuration

See FunctionConfiguration

function_configuration_environment

See FunctionConfigurationEnvironment

function_default_config

See FunctionDefaultConfig

function_default_execution_config

See FunctionDefaultExecutionConfig

function_definition_version

See FunctionDefinitionVersion

function_execution_config

See FunctionExecutionConfig

function_run_as_config

See FunctionRunAsConfig

group_certificate_authority_properties

See GroupCertificateAuthorityProperties

group_information

See GroupInformation

group_owner_setting

See GroupOwnerSetting

group_version

See GroupVersion

local_device_resource_data

See LocalDeviceResourceData

local_volume_resource_data

See LocalVolumeResourceData

logger

See Logger

logger_definition_version

See LoggerDefinitionVersion

resource

See Resource

resource_access_policy

See ResourceAccessPolicy

resource_data_container

See ResourceDataContainer

resource_definition_version

See ResourceDefinitionVersion

resource_download_owner_setting

See ResourceDownloadOwnerSetting

runtime_configuration

See RuntimeConfiguration

s3_machine_learning_model_resource_data

See S3MachineLearningModelResourceData

sage_maker_machine_learning_model_resource_data

See SageMakerMachineLearningModelResourceData

secrets_manager_secret_resource_data

See SecretsManagerSecretResourceData

subscription

See Subscription

subscription_definition_version

See SubscriptionDefinitionVersion

telemetry_configuration

See TelemetryConfiguration

telemetry_configuration_update

See TelemetryConfigurationUpdate

version_information

See VersionInformation

Structs

BulkDeployment

Information about a bulk deployment. You cannot start a new bulk deployment while another one is still running or in a non-terminal state.

BulkDeploymentMetrics

Relevant metrics on input records processed during bulk deployment.

BulkDeploymentResult

Information about an individual group deployment in a bulk deployment operation.

ConnectivityInfo

Information about a Greengrass core’s connectivity.

Connector

Information about a connector. Connectors run on the Greengrass core and contain built-in integration with local infrastructure, device protocols, AWS, and other cloud services.

ConnectorDefinitionVersion

Information about the connector definition version, which is a container for connectors.

Core

Information about a core.

CoreDefinitionVersion

Information about a core definition version.

DefinitionInformation

Information about a definition.

Deployment

Information about a deployment.

Device

Information about a device.

DeviceDefinitionVersion

Information about a device definition version.

ErrorDetail

Details about the error.

Function

Information about a Lambda function.

FunctionConfiguration

The configuration of the Lambda function.

FunctionConfigurationEnvironment

The environment configuration of the function.

FunctionDefaultConfig

The default configuration that applies to all Lambda functions in the group. Individual Lambda functions can override these settings.

FunctionDefaultExecutionConfig

Configuration information that specifies how a Lambda function runs.

FunctionDefinitionVersion

Information about a function definition version.

FunctionExecutionConfig

Configuration information that specifies how a Lambda function runs.

FunctionRunAsConfig

Specifies the user and group whose permissions are used when running the Lambda function. You can specify one or both values to override the default values. We recommend that you avoid running as root unless absolutely necessary to minimize the risk of unintended changes or malicious attacks. To run as root, you must set ’‘IsolationMode’’ to ’‘NoContainer’’ and update config.json in ’‘greengrass-root/config’’ to set ’‘allowFunctionsToRunAsRoot’’ to ’‘yes’’.

GroupCertificateAuthorityProperties

Information about a certificate authority for a group.

GroupInformation

Information about a group.

GroupOwnerSetting

Group owner related settings for local resources.

GroupVersion

Information about a group version.

LocalDeviceResourceData

Attributes that define a local device resource.

LocalVolumeResourceData

Attributes that define a local volume resource.

Logger

Information about a logger

LoggerDefinitionVersion

Information about a logger definition version.

Resource

Information about a resource.

ResourceAccessPolicy

A policy used by the function to access a resource.

ResourceDataContainer

A container for resource data. The container takes only one of the following supported resource data types: ’‘LocalDeviceResourceData’’, ’‘LocalVolumeResourceData’’, ’‘SageMakerMachineLearningModelResourceData’’, ’‘S3MachineLearningModelResourceData’’, ’‘SecretsManagerSecretResourceData’’.

ResourceDefinitionVersion

Information about a resource definition version.

ResourceDownloadOwnerSetting

The owner setting for downloaded machine learning resources.

RuntimeConfiguration

Runtime configuration for a thing.

S3MachineLearningModelResourceData

Attributes that define an Amazon S3 machine learning resource.

SageMakerMachineLearningModelResourceData

Attributes that define an Amazon SageMaker machine learning resource.

SecretsManagerSecretResourceData

Attributes that define a secret resource, which references a secret from AWS Secrets Manager. AWS IoT Greengrass stores a local, encrypted copy of the secret on the Greengrass core, where it can be securely accessed by connectors and Lambda functions.

Subscription

Information about a subscription.

SubscriptionDefinitionVersion

Information about a subscription definition version.

TelemetryConfiguration

Configuration settings for running telemetry.

TelemetryConfigurationUpdate

Configuration settings for running telemetry.

VersionInformation

Information about a version.

Enums

BulkDeploymentStatus

The current status of the bulk deployment.

ConfigurationSyncStatus

DeploymentType

The type of deployment. When used for ’‘CreateDeployment’’, only ’‘NewDeployment’’ and ’‘Redeployment’’ are valid.

EncodingType

FunctionIsolationMode

Specifies whether the Lambda function runs in a Greengrass container (default) or without containerization. Unless your scenario requires that you run without containerization, we recommend that you run in a Greengrass container. Omit this value to run the Lambda function with the default containerization for the group.

LoggerComponent

LoggerLevel

LoggerType

Permission

The type of permission a function has to access a resource.

SoftwareToUpdate

The piece of software on the Greengrass core that will be updated.

Telemetry

UpdateAgentLogLevel

The minimum level of log statements that should be logged by the OTA Agent during an update.

UpdateTargetsArchitecture

The architecture of the cores which are the targets of an update.

UpdateTargetsOperatingSystem

The operating system of the cores which are the targets of an update.