Struct aws_sdk_greengrass::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for Greengrass.

This client allows ergonomic access to a Greengrass-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_role_to_group(&self) -> AssociateRoleToGroup<C, M, R>

Constructs a fluent builder for the AssociateRoleToGroup operation.

See AssociateRoleToGroup for more information about the operation and its arguments.

pub fn associate_service_role_to_account(
&self
) -> AssociateServiceRoleToAccount<C, M, R>

Constructs a fluent builder for the AssociateServiceRoleToAccount operation.

See AssociateServiceRoleToAccount for more information about the operation and its arguments.

pub fn create_connector_definition(&self) -> CreateConnectorDefinition<C, M, R>

Constructs a fluent builder for the CreateConnectorDefinition operation.

See CreateConnectorDefinition for more information about the operation and its arguments.

pub fn create_connector_definition_version(
&self
) -> CreateConnectorDefinitionVersion<C, M, R>

Constructs a fluent builder for the CreateConnectorDefinitionVersion operation.

See CreateConnectorDefinitionVersion for more information about the operation and its arguments.

pub fn create_core_definition(&self) -> CreateCoreDefinition<C, M, R>

Constructs a fluent builder for the CreateCoreDefinition operation.

See CreateCoreDefinition for more information about the operation and its arguments.

pub fn create_core_definition_version(
&self
) -> CreateCoreDefinitionVersion<C, M, R>

Constructs a fluent builder for the CreateCoreDefinitionVersion operation.

See CreateCoreDefinitionVersion for more information about the operation and its arguments.

pub fn create_deployment(&self) -> CreateDeployment<C, M, R>

Constructs a fluent builder for the CreateDeployment operation.

See CreateDeployment for more information about the operation and its arguments.

pub fn create_device_definition(&self) -> CreateDeviceDefinition<C, M, R>

Constructs a fluent builder for the CreateDeviceDefinition operation.

See CreateDeviceDefinition for more information about the operation and its arguments.

pub fn create_device_definition_version(
&self
) -> CreateDeviceDefinitionVersion<C, M, R>

Constructs a fluent builder for the CreateDeviceDefinitionVersion operation.

See CreateDeviceDefinitionVersion for more information about the operation and its arguments.

pub fn create_function_definition(&self) -> CreateFunctionDefinition<C, M, R>

Constructs a fluent builder for the CreateFunctionDefinition operation.

See CreateFunctionDefinition for more information about the operation and its arguments.

pub fn create_function_definition_version(
&self
) -> CreateFunctionDefinitionVersion<C, M, R>

Constructs a fluent builder for the CreateFunctionDefinitionVersion operation.

See CreateFunctionDefinitionVersion for more information about the operation and its arguments.

pub fn create_group(&self) -> CreateGroup<C, M, R>

Constructs a fluent builder for the CreateGroup operation.

See CreateGroup for more information about the operation and its arguments.

pub fn create_group_certificate_authority(
&self
) -> CreateGroupCertificateAuthority<C, M, R>

Constructs a fluent builder for the CreateGroupCertificateAuthority operation.

See CreateGroupCertificateAuthority for more information about the operation and its arguments.

pub fn create_group_version(&self) -> CreateGroupVersion<C, M, R>

Constructs a fluent builder for the CreateGroupVersion operation.

See CreateGroupVersion for more information about the operation and its arguments.

pub fn create_logger_definition(&self) -> CreateLoggerDefinition<C, M, R>

Constructs a fluent builder for the CreateLoggerDefinition operation.

See CreateLoggerDefinition for more information about the operation and its arguments.

pub fn create_logger_definition_version(
&self
) -> CreateLoggerDefinitionVersion<C, M, R>

Constructs a fluent builder for the CreateLoggerDefinitionVersion operation.

See CreateLoggerDefinitionVersion for more information about the operation and its arguments.

pub fn create_resource_definition(&self) -> CreateResourceDefinition<C, M, R>

Constructs a fluent builder for the CreateResourceDefinition operation.

See CreateResourceDefinition for more information about the operation and its arguments.

pub fn create_resource_definition_version(
&self
) -> CreateResourceDefinitionVersion<C, M, R>

Constructs a fluent builder for the CreateResourceDefinitionVersion operation.

See CreateResourceDefinitionVersion for more information about the operation and its arguments.

pub fn create_software_update_job(&self) -> CreateSoftwareUpdateJob<C, M, R>

Constructs a fluent builder for the CreateSoftwareUpdateJob operation.

See CreateSoftwareUpdateJob for more information about the operation and its arguments.

pub fn create_subscription_definition(
&self
) -> CreateSubscriptionDefinition<C, M, R>

Constructs a fluent builder for the CreateSubscriptionDefinition operation.

See CreateSubscriptionDefinition for more information about the operation and its arguments.

pub fn create_subscription_definition_version(
&self
) -> CreateSubscriptionDefinitionVersion<C, M, R>

Constructs a fluent builder for the CreateSubscriptionDefinitionVersion operation.

See CreateSubscriptionDefinitionVersion for more information about the operation and its arguments.

pub fn delete_connector_definition(&self) -> DeleteConnectorDefinition<C, M, R>

Constructs a fluent builder for the DeleteConnectorDefinition operation.

See DeleteConnectorDefinition for more information about the operation and its arguments.

pub fn delete_core_definition(&self) -> DeleteCoreDefinition<C, M, R>

Constructs a fluent builder for the DeleteCoreDefinition operation.

See DeleteCoreDefinition for more information about the operation and its arguments.

pub fn delete_device_definition(&self) -> DeleteDeviceDefinition<C, M, R>

Constructs a fluent builder for the DeleteDeviceDefinition operation.

See DeleteDeviceDefinition for more information about the operation and its arguments.

pub fn delete_function_definition(&self) -> DeleteFunctionDefinition<C, M, R>

Constructs a fluent builder for the DeleteFunctionDefinition operation.

See DeleteFunctionDefinition for more information about the operation and its arguments.

pub fn delete_group(&self) -> DeleteGroup<C, M, R>

Constructs a fluent builder for the DeleteGroup operation.

See DeleteGroup for more information about the operation and its arguments.

pub fn delete_logger_definition(&self) -> DeleteLoggerDefinition<C, M, R>

Constructs a fluent builder for the DeleteLoggerDefinition operation.

See DeleteLoggerDefinition for more information about the operation and its arguments.

pub fn delete_resource_definition(&self) -> DeleteResourceDefinition<C, M, R>

Constructs a fluent builder for the DeleteResourceDefinition operation.

See DeleteResourceDefinition for more information about the operation and its arguments.

pub fn delete_subscription_definition(
&self
) -> DeleteSubscriptionDefinition<C, M, R>

Constructs a fluent builder for the DeleteSubscriptionDefinition operation.

See DeleteSubscriptionDefinition for more information about the operation and its arguments.

pub fn disassociate_role_from_group(&self) -> DisassociateRoleFromGroup<C, M, R>

Constructs a fluent builder for the DisassociateRoleFromGroup operation.

See DisassociateRoleFromGroup for more information about the operation and its arguments.

pub fn disassociate_service_role_from_account(
&self
) -> DisassociateServiceRoleFromAccount<C, M, R>

Constructs a fluent builder for the DisassociateServiceRoleFromAccount operation.

See DisassociateServiceRoleFromAccount for more information about the operation and its arguments.

pub fn get_associated_role(&self) -> GetAssociatedRole<C, M, R>

Constructs a fluent builder for the GetAssociatedRole operation.

See GetAssociatedRole for more information about the operation and its arguments.

pub fn get_bulk_deployment_status(&self) -> GetBulkDeploymentStatus<C, M, R>

Constructs a fluent builder for the GetBulkDeploymentStatus operation.

See GetBulkDeploymentStatus for more information about the operation and its arguments.

pub fn get_connectivity_info(&self) -> GetConnectivityInfo<C, M, R>

Constructs a fluent builder for the GetConnectivityInfo operation.

See GetConnectivityInfo for more information about the operation and its arguments.

pub fn get_connector_definition(&self) -> GetConnectorDefinition<C, M, R>

Constructs a fluent builder for the GetConnectorDefinition operation.

See GetConnectorDefinition for more information about the operation and its arguments.

pub fn get_connector_definition_version(
&self
) -> GetConnectorDefinitionVersion<C, M, R>

Constructs a fluent builder for the GetConnectorDefinitionVersion operation.

See GetConnectorDefinitionVersion for more information about the operation and its arguments.

pub fn get_core_definition(&self) -> GetCoreDefinition<C, M, R>

Constructs a fluent builder for the GetCoreDefinition operation.

See GetCoreDefinition for more information about the operation and its arguments.

pub fn get_core_definition_version(&self) -> GetCoreDefinitionVersion<C, M, R>

Constructs a fluent builder for the GetCoreDefinitionVersion operation.

See GetCoreDefinitionVersion for more information about the operation and its arguments.

pub fn get_deployment_status(&self) -> GetDeploymentStatus<C, M, R>

Constructs a fluent builder for the GetDeploymentStatus operation.

See GetDeploymentStatus for more information about the operation and its arguments.

pub fn get_device_definition(&self) -> GetDeviceDefinition<C, M, R>

Constructs a fluent builder for the GetDeviceDefinition operation.

See GetDeviceDefinition for more information about the operation and its arguments.

pub fn get_device_definition_version(
&self
) -> GetDeviceDefinitionVersion<C, M, R>

Constructs a fluent builder for the GetDeviceDefinitionVersion operation.

See GetDeviceDefinitionVersion for more information about the operation and its arguments.

pub fn get_function_definition(&self) -> GetFunctionDefinition<C, M, R>

Constructs a fluent builder for the GetFunctionDefinition operation.

See GetFunctionDefinition for more information about the operation and its arguments.

pub fn get_function_definition_version(
&self
) -> GetFunctionDefinitionVersion<C, M, R>

Constructs a fluent builder for the GetFunctionDefinitionVersion operation.

See GetFunctionDefinitionVersion for more information about the operation and its arguments.

pub fn get_group(&self) -> GetGroup<C, M, R>

Constructs a fluent builder for the GetGroup operation.

See GetGroup for more information about the operation and its arguments.

pub fn get_group_certificate_authority(
&self
) -> GetGroupCertificateAuthority<C, M, R>

Constructs a fluent builder for the GetGroupCertificateAuthority operation.

See GetGroupCertificateAuthority for more information about the operation and its arguments.

pub fn get_group_certificate_configuration(
&self
) -> GetGroupCertificateConfiguration<C, M, R>

Constructs a fluent builder for the GetGroupCertificateConfiguration operation.

See GetGroupCertificateConfiguration for more information about the operation and its arguments.

pub fn get_group_version(&self) -> GetGroupVersion<C, M, R>

Constructs a fluent builder for the GetGroupVersion operation.

See GetGroupVersion for more information about the operation and its arguments.

pub fn get_logger_definition(&self) -> GetLoggerDefinition<C, M, R>

Constructs a fluent builder for the GetLoggerDefinition operation.

See GetLoggerDefinition for more information about the operation and its arguments.

pub fn get_logger_definition_version(
&self
) -> GetLoggerDefinitionVersion<C, M, R>

Constructs a fluent builder for the GetLoggerDefinitionVersion operation.

See GetLoggerDefinitionVersion for more information about the operation and its arguments.

pub fn get_resource_definition(&self) -> GetResourceDefinition<C, M, R>

Constructs a fluent builder for the GetResourceDefinition operation.

See GetResourceDefinition for more information about the operation and its arguments.

pub fn get_resource_definition_version(
&self
) -> GetResourceDefinitionVersion<C, M, R>

Constructs a fluent builder for the GetResourceDefinitionVersion operation.

See GetResourceDefinitionVersion for more information about the operation and its arguments.

pub fn get_service_role_for_account(&self) -> GetServiceRoleForAccount<C, M, R>

Constructs a fluent builder for the GetServiceRoleForAccount operation.

See GetServiceRoleForAccount for more information about the operation and its arguments.

pub fn get_subscription_definition(&self) -> GetSubscriptionDefinition<C, M, R>

Constructs a fluent builder for the GetSubscriptionDefinition operation.

See GetSubscriptionDefinition for more information about the operation and its arguments.

pub fn get_subscription_definition_version(
&self
) -> GetSubscriptionDefinitionVersion<C, M, R>

Constructs a fluent builder for the GetSubscriptionDefinitionVersion operation.

See GetSubscriptionDefinitionVersion for more information about the operation and its arguments.

pub fn get_thing_runtime_configuration(
&self
) -> GetThingRuntimeConfiguration<C, M, R>

Constructs a fluent builder for the GetThingRuntimeConfiguration operation.

See GetThingRuntimeConfiguration for more information about the operation and its arguments.

pub fn list_bulk_deployment_detailed_reports(
&self
) -> ListBulkDeploymentDetailedReports<C, M, R>

Constructs a fluent builder for the ListBulkDeploymentDetailedReports operation.

See ListBulkDeploymentDetailedReports for more information about the operation and its arguments.

pub fn list_bulk_deployments(&self) -> ListBulkDeployments<C, M, R>

Constructs a fluent builder for the ListBulkDeployments operation.

See ListBulkDeployments for more information about the operation and its arguments.

pub fn list_connector_definitions(&self) -> ListConnectorDefinitions<C, M, R>

Constructs a fluent builder for the ListConnectorDefinitions operation.

See ListConnectorDefinitions for more information about the operation and its arguments.

pub fn list_connector_definition_versions(
&self
) -> ListConnectorDefinitionVersions<C, M, R>

Constructs a fluent builder for the ListConnectorDefinitionVersions operation.

See ListConnectorDefinitionVersions for more information about the operation and its arguments.

pub fn list_core_definitions(&self) -> ListCoreDefinitions<C, M, R>

Constructs a fluent builder for the ListCoreDefinitions operation.

See ListCoreDefinitions for more information about the operation and its arguments.

pub fn list_core_definition_versions(
&self
) -> ListCoreDefinitionVersions<C, M, R>

Constructs a fluent builder for the ListCoreDefinitionVersions operation.

See ListCoreDefinitionVersions for more information about the operation and its arguments.

pub fn list_deployments(&self) -> ListDeployments<C, M, R>

Constructs a fluent builder for the ListDeployments operation.

See ListDeployments for more information about the operation and its arguments.

pub fn list_device_definitions(&self) -> ListDeviceDefinitions<C, M, R>

Constructs a fluent builder for the ListDeviceDefinitions operation.

See ListDeviceDefinitions for more information about the operation and its arguments.

pub fn list_device_definition_versions(
&self
) -> ListDeviceDefinitionVersions<C, M, R>

Constructs a fluent builder for the ListDeviceDefinitionVersions operation.

See ListDeviceDefinitionVersions for more information about the operation and its arguments.

pub fn list_function_definitions(&self) -> ListFunctionDefinitions<C, M, R>

Constructs a fluent builder for the ListFunctionDefinitions operation.

See ListFunctionDefinitions for more information about the operation and its arguments.

pub fn list_function_definition_versions(
&self
) -> ListFunctionDefinitionVersions<C, M, R>

Constructs a fluent builder for the ListFunctionDefinitionVersions operation.

See ListFunctionDefinitionVersions for more information about the operation and its arguments.

pub fn list_group_certificate_authorities(
&self
) -> ListGroupCertificateAuthorities<C, M, R>

Constructs a fluent builder for the ListGroupCertificateAuthorities operation.

See ListGroupCertificateAuthorities for more information about the operation and its arguments.

pub fn list_groups(&self) -> ListGroups<C, M, R>

Constructs a fluent builder for the ListGroups operation.

See ListGroups for more information about the operation and its arguments.

pub fn list_group_versions(&self) -> ListGroupVersions<C, M, R>

Constructs a fluent builder for the ListGroupVersions operation.

See ListGroupVersions for more information about the operation and its arguments.

pub fn list_logger_definitions(&self) -> ListLoggerDefinitions<C, M, R>

Constructs a fluent builder for the ListLoggerDefinitions operation.

See ListLoggerDefinitions for more information about the operation and its arguments.

pub fn list_logger_definition_versions(
&self
) -> ListLoggerDefinitionVersions<C, M, R>

Constructs a fluent builder for the ListLoggerDefinitionVersions operation.

See ListLoggerDefinitionVersions for more information about the operation and its arguments.

pub fn list_resource_definitions(&self) -> ListResourceDefinitions<C, M, R>

Constructs a fluent builder for the ListResourceDefinitions operation.

See ListResourceDefinitions for more information about the operation and its arguments.

pub fn list_resource_definition_versions(
&self
) -> ListResourceDefinitionVersions<C, M, R>

Constructs a fluent builder for the ListResourceDefinitionVersions operation.

See ListResourceDefinitionVersions for more information about the operation and its arguments.

pub fn list_subscription_definitions(
&self
) -> ListSubscriptionDefinitions<C, M, R>

Constructs a fluent builder for the ListSubscriptionDefinitions operation.

See ListSubscriptionDefinitions for more information about the operation and its arguments.

pub fn list_subscription_definition_versions(
&self
) -> ListSubscriptionDefinitionVersions<C, M, R>

Constructs a fluent builder for the ListSubscriptionDefinitionVersions operation.

See ListSubscriptionDefinitionVersions for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn reset_deployments(&self) -> ResetDeployments<C, M, R>

Constructs a fluent builder for the ResetDeployments operation.

See ResetDeployments for more information about the operation and its arguments.

pub fn start_bulk_deployment(&self) -> StartBulkDeployment<C, M, R>

Constructs a fluent builder for the StartBulkDeployment operation.

See StartBulkDeployment for more information about the operation and its arguments.

pub fn stop_bulk_deployment(&self) -> StopBulkDeployment<C, M, R>

Constructs a fluent builder for the StopBulkDeployment operation.

See StopBulkDeployment for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_connectivity_info(&self) -> UpdateConnectivityInfo<C, M, R>

Constructs a fluent builder for the UpdateConnectivityInfo operation.

See UpdateConnectivityInfo for more information about the operation and its arguments.

pub fn update_connector_definition(&self) -> UpdateConnectorDefinition<C, M, R>

Constructs a fluent builder for the UpdateConnectorDefinition operation.

See UpdateConnectorDefinition for more information about the operation and its arguments.

pub fn update_core_definition(&self) -> UpdateCoreDefinition<C, M, R>

Constructs a fluent builder for the UpdateCoreDefinition operation.

See UpdateCoreDefinition for more information about the operation and its arguments.

pub fn update_device_definition(&self) -> UpdateDeviceDefinition<C, M, R>

Constructs a fluent builder for the UpdateDeviceDefinition operation.

See UpdateDeviceDefinition for more information about the operation and its arguments.

pub fn update_function_definition(&self) -> UpdateFunctionDefinition<C, M, R>

Constructs a fluent builder for the UpdateFunctionDefinition operation.

See UpdateFunctionDefinition for more information about the operation and its arguments.

pub fn update_group(&self) -> UpdateGroup<C, M, R>

Constructs a fluent builder for the UpdateGroup operation.

See UpdateGroup for more information about the operation and its arguments.

pub fn update_group_certificate_configuration(
&self
) -> UpdateGroupCertificateConfiguration<C, M, R>

Constructs a fluent builder for the UpdateGroupCertificateConfiguration operation.

See UpdateGroupCertificateConfiguration for more information about the operation and its arguments.

pub fn update_logger_definition(&self) -> UpdateLoggerDefinition<C, M, R>

Constructs a fluent builder for the UpdateLoggerDefinition operation.

See UpdateLoggerDefinition for more information about the operation and its arguments.

pub fn update_resource_definition(&self) -> UpdateResourceDefinition<C, M, R>

Constructs a fluent builder for the UpdateResourceDefinition operation.

See UpdateResourceDefinition for more information about the operation and its arguments.

pub fn update_subscription_definition(
&self
) -> UpdateSubscriptionDefinition<C, M, R>

Constructs a fluent builder for the UpdateSubscriptionDefinition operation.

See UpdateSubscriptionDefinition for more information about the operation and its arguments.

pub fn update_thing_runtime_configuration(
&self
) -> UpdateThingRuntimeConfiguration<C, M, R>

Constructs a fluent builder for the UpdateThingRuntimeConfiguration operation.

See UpdateThingRuntimeConfiguration for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.